% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json.R
\name{prepare.elements}
\alias{prepare.elements}
\title{Prepare json elements}
\usage{
prepare.elements(json, depth = 1)
}
\arguments{
\item{json}{list which elements are parsed into json format.}
}
\value{
Returns json format of each element in the list.
}
\description{
Get and format all the elements in the list into a json format.
It takes in a list and for each element produces the json format of that element.
}
