% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json.R
\name{prepare.json}
\alias{prepare.json}
\title{Put together the final json}
\usage{
prepare.json(json)
}
\arguments{
\item{json}{list that is to be parsed into json.}
}
\description{
Take list elements, that have been parsed into json format and combine them into single json element.
}
