% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read.json}
\alias{read.json}
\title{Read json in as list}
\usage{
read.json(file = "")
}
\arguments{
\item{file}{the name of the file, where the json is read from.}
}
\description{
Read in a json file and parse it into a R list.
}
