% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_json.R
\name{write.json}
\alias{write.json}
\title{Write list as json}
\usage{
write.json(to_json, file = "new_file.json")
}
\arguments{
\item{to_json}{list that is to be written out as json.}

\item{file}{the name of the file, where the json is written. By default it is written in the current working directory as 'new_file.json'.}
}
\description{
Write given list to given file in json format. Structures such as data.frame and data.table work as well (also inside the list).
}
\examples{
write.json(lst) # The argument value of lst is in examples/example_list.txt. The result can also be found at examples/example_list_written_out.json
write.json(iris)
write.json(cars)
}
