% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json.R
\name{read.json}
\alias{read.json}
\title{Read json in as a list}
\usage{
read.json(file = "", url = "")
}
\arguments{
\item{file}{the name of the file, where the json is read from.}

\item{url}{the url of the json to read.}
}
\description{
Read in a json file and parse it into a R list. If both parameters (file,url) are provided, then url value is used.
}
\examples{
read.json('examples/example_json_self.json') # Reads in an example json as list from examples/.
read.json('examples/example_json_self_written.json') # Reads in an example json as list from examples/.
read.json('examples/dnd_example.json') # Reads in an example json as list from examples/.
read.json(url = 'https://www.dnd5eapi.co/api/spells/acid-arrow/' ) # Reads in json from given url as list.
}
